DSL using 'ai' instead of AI
