+++ examples/scala/src/main/scala/com/xebia/functional/xef/scala/auto/Animal.scala
@@ -13,7 +13,7 @@ private final case class Invention(name: String, inventor: String, year: Int, pu
 private final case class Story(animal: Animal, invention: Invention, story: String) derives ScalaSerialDescriptor, Decoder
 
 @main def runAnimal: Unit =
-  AI {
+  ai {
     val animal: Animal = prompt("A unique animal species")
     val invention: Invention = prompt("A groundbreaking invention from the 20th century.")
