+++ examples/scala/src/main/scala/com/xebia/functional/xef/scala/auto/Book.scala
@@ -8,5 +8,5 @@ import io.circe.Decoder
 private final case class Book(name: String, author: String, summary: String) derives ScalaSerialDescriptor, Decoder
 
 @main def runBook: Unit =
-  val book = AI(prompt[Book]("To Kill a Mockingbird by Harper Lee summary."))
+  val book = ai(prompt[Book]("To Kill a Mockingbird by Harper Lee summary."))
   println(s"To Kill a Mockingbird summary:\n ${book.summary}")
