+++ scala/src/main/scala/com/xebia/functional/xef/scala/auto/AIScope.scala
@@ -1,0 +1,7 @@
+package com.xebia.functional.xef.scala.auto
+
+import com.xebia.functional.xef.auto.{AIException, AIKt, AIScope as KtAIScope, Agent as KtAgent}
+
+final case class AIScope(kt: KtAIScope)
+private object AIScope:
+  def fromCore(coreAIScope: KtAIScope): AIScope = new AIScope(coreAIScope)
